add a unique 64 bit number in each AST node

+++ scilab/modules/ast/src/cpp/visitor_common.cpp
@@ -25,7 +25,7 @@
 
 using namespace types;
 
-int ast::Ast::globalNodeNumber = 0;
+size_t ast::Ast::globalNodeNumber = 0;
 static bool isIntTrue(InternalType* _pIT);
 
 bool bConditionState(types::InternalType *_pITResult)
